merge in FSL BSP files for initializing I2S
Basic I2S initialization done. Mostly the FSL merge was to
be able to handle MCLK PLL calculations cleanly, as they had
all the constants pre-defined. It's not clear if this was
a good idea, because the next commit is about to pull in a gob
of files from the freescale clock tree just to handle all of
their standard HAL BS.