I2S kind of working-ish
I2S runs, at the right frequency, with the right data format
on the wire. The RX buffer fills, and copies to a back buffer.
The data kind of looks okay, although I need to sift through
the settings to rectify the weirdo format that the microphone
outputs versus what the Kinetis can handle; may require some
bit packing and oscope work to figure out which phase the
receiver is on and which portions of the Rx frame are good.

That comes later, though. At the moment hitting any key on
the keyboard (or any serial traffic, including internal print
messages) will cause the system to hang on an SV8. So there's
a locking issue.

Now pondering if the problem is that the ISR is too long,
or if the locking primtives I'm trying to use aren't the
right ones. I'm guessing it's the latter?