clean up I2S driver, fix minor bugs
- clean up Freescale's crappy FIFO drain code. Can't change
word sizes anymore, but with this chip, it's not possible so
things will run better.

- double MCLK speed yet again (so it's 4x BCLK speed). The
datasheet doesn't specify a minimum MCLK:BCLK ratio, but I
suspect strongly that it should be at least 4x over BCLK.
I was seeing some weird data dropout with a lower ratio? I
think it's now at a sweet spot and at one of the "standard"
frequencies for MCLK.

- masked out unused stereo channel from entering the FIFO,
effectively reducing # of interrupts in half.

- added some debugging commands in the command shell.