fix data sample type and alignment error
data sample type adjusted to int32_t (to match 2's compliment
format of the data read in).

the dumping function now reflects the correct data type. Forgot
to update orchard-i2s.h with the data type and so was dumping
16-bit samples on 32-bit data causing mysterious and frustrating
debugging problems. GDB to the rescue, printf for the fail!